agent/requirements.txt
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-naver>=0.1.0
python-dotenv>=1.0.0 